Get all componetes working... i think
