+++ app/components/CheckInModal.js
@@ -1,11 +1,18 @@
 import { useState } from "react";
-import { motion } from "framer-motion";
 import GuestBook from "./GuestBook";
 import Ticket from "./Ticket";
 
 export default function CheckInModal({ checkInRef }) {
   const [isExpanded, setIsExpanded] = useState(false);
-  // const [page, setPage] = useState(1);
+  const [page, setPage] = useState(1);
+
+  const toggleBook = () => {
+    setIsExpanded((prev) => {
+      const newState = !prev;
+      setPage(newState ? 2 : 1); // Set page to 2 when expanded, else set to 1
+      return newState;
+    });
+  };
 
   return (
     <dialog
@@ -22,7 +29,7 @@ export default function CheckInModal({ checkInRef }) {
 
       {/* Main Content */}
       <div
-        className={`pb-4 max-w-5xl m-auto gap-4 h-[80vh] overflow-hidden flex px-8 flex-col lg:flex-row transition-transform duration-[10000] ${
+        className={`max-w-5xl m-auto gap-4 h-[80vh] overflow-hidden flex px-8 flex-col lg:flex-row transition-transform duration-500 ${
           isExpanded ? "-translate-x-[0%]" : ""
         }`}
       >
@@ -34,13 +41,8 @@ export default function CheckInModal({ checkInRef }) {
         )}
 
         {/* GuestBook Component */}
-
-        <div
-          className={` z-50 transition-all duration-500 relative ${
-            isExpanded ? "w-full" : "w-full"
-          }`}
-        >
-          <GuestBook isExpanded={isExpanded} setIsExpanded={setIsExpanded} />
+        <div className={`z-50 transition-all duration-500 relative w-full`}>
+          <GuestBook page={page} setPage={setPage} toggleBook={toggleBook} />
         </div>
       </div>
 
@@ -51,16 +53,10 @@ export default function CheckInModal({ checkInRef }) {
             process.env.NEXT_PUBLIC_API_BASE_PATH || ""
           }/images/textures/wood.jpg')`,
         }}
-        className="w-full flex justify-center"
+        className="w-full h-full flex justify-center gap-10"
       >
-        {/* <button
-          className="z-50 h-min p-6 mt-4 skeu w-[200px]"
-        >
-          Surprise me 🎊
-        </button> */}
-
         {/* Plaque */}
-        <div className="relative h-min w-max p-4 bg-gradient-to-br from-orange-300 to-orange-500 border-2 border-orange-400 rounded-md shadow-sm">
+        <div className="mt-4 relative h-min w-max p-4 bg-gradient-to-br from-orange-300 to-orange-500 border-2 border-orange-400 rounded-md shadow-sm">
           <h3 className="font-bold px-4 text-2xl text-center uppercase font-serif text-yellow-800 drop-shadow-[0_1px_0_rgba(255,255,255,0.8)] tracking-widest">
             Ticket Desk
           </h3>
@@ -79,10 +75,6 @@ export default function CheckInModal({ checkInRef }) {
             <div className="absolute bottom-2 right-2 w-2 h-2 rounded-full bg-orange-700 border border-orange-900 shadow-[inset_0px_2px_3px_rgba(255,255,255,0.6)]"></div>
           </div>
         </div>
-
-        {/* <button onClick={() => toggleBook()} className="z-50 h-min p-6 mt-4 skeu w-[200px]">
-          {isExpanded ? "Close Guest Book" : "Open Guest Book"}
-        </button> */}
       </div>
     </dialog>
   );
